Allow unit test timeouts to be specified using nanosecond precision
Previously, test timeouts were limited to second precision. The reason
was the call used to determine test timeouts, alarm, only supported
second precision. With this change, three more functions from librt
are used set a timer to fire with nanosecond precision. For systems
without librt, an implementation of these functions is in libcompat
which fall back on using alarm.

git-svn-id: svn+ssh://svn.code.sf.net/p/check/code/trunk@679 64e312b2-a51f-0410-8e61-82d0ca0eb02a