modify test_log_output.sh to compare output based on HAVE_FORK
When run in Windows, the output differs slightly from when it is run
under *NIX due to line endings. *NIX uses \n, whereas Windows
uses \r\n. To account for this, printf is used to generate the
Windows expected string.

Additionally, the test_exit test was removed, as it is not valid
if fork is unavailable.

git-svn-id: svn+ssh://svn.code.sf.net/p/check/code/trunk@782 64e312b2-a51f-0410-8e61-82d0ca0eb02a