Disable checked fixtures if fork is unavailable
checked fixtures allow setup and teardown functions to be run in the
same process space as tests. If a fixture fails, the error is caught
and reported as a failure for the associated test.

If fork is not used, then the guarantee that checked fixtures try to
provide is not possible. Further, using checked fixtures without fork can
cause issues. For example, without fork if a checked teardown
calls ck_assert() which fails, this results in longjmp being called,
which results in the teardowns being called again. An infinate loop
results.

git-svn-id: svn+ssh://svn.code.sf.net/p/check/code/trunk@795 64e312b2-a51f-0410-8e61-82d0ca0eb02a