web: change layout of landing page
Instead of a plain black/white page with text, the new layout
contains some color, a side pane with links and info on the
latest release, and some links.

All data related to the page is in the new subfolder "web".

The layout was from Open Source Web Design (oswd.org), and is
the NewsPortal layout by "Designs by Darren".

git-svn-id: svn+ssh://svn.code.sf.net/p/check/code/trunk@864 64e312b2-a51f-0410-8e61-82d0ca0eb02a