Add #define for disabling unit tests that intentionally leak memory
Some tests intentionally leak memory. Usually this is not a concern.
However, we want to run some of our tests against valgrind, to make
sure that Check itself is not leaking any memory. To enable this,
the tests that intentionally leak memory will be conditionally
compiled out when building a unit tests program for use with
valgrind.

git-svn-id: svn+ssh://svn.code.sf.net/p/check/code/trunk@914 64e312b2-a51f-0410-8e61-82d0ca0eb02a