check_check_sub: sleep during test_eternal
For some reason on Linux (maybe other systems) if
check_check is canceled (e.g. Ctrl+C) when the test_eternal
is running, check_check will still run in the background
on this test. Because the test is in a tight loop,
it eats lots of CPU. Replacing the busy loop with the sleep
does not fix the problem, but at least when it happens
it will not take lots of CPU.

git-svn-id: svn+ssh://svn.code.sf.net/p/check/code/trunk@989 64e312b2-a51f-0410-8e61-82d0ca0eb02a