simplify path and file name handling and honor const rules
the path and file name variables are not const, but were assigned
const strings. However, both variables could be reassigned
non const data, and the path_name could be allocated data.
To prevent this mix, they are both non-const and initially NULL.
If they are NULL during the print, they are replaced before printing.
Further, the path_name is now always free'd at the end, as it is
either NULL or allocated memory.

git-svn-id: svn+ssh://svn.code.sf.net/p/check/code/trunk@1066 64e312b2-a51f-0410-8e61-82d0ca0eb02a