fix formatting for ck_assert_(u)int_* calls if using % operator
If an expression is used in any of the ck_assert_(u)int_* calls
that uses the % operator, for example:

ck_assert_int_eq(fs.write.slot % fs.slots_per_sector, i+1);

the result was the "% f" would be interpreted as a format string.

This fix prevents the issue by passing the entire expression as a
string, and feeding the underlying printf a "%s" argument instead.

git-svn-id: svn+ssh://svn.code.sf.net/p/check/code/trunk@1126 64e312b2-a51f-0410-8e61-82d0ca0eb02a