Use env to find shell to use
On at least one platform (Solaris), the shell located at /bin/sh is
not a compliant shell. Using env as a form of redirection to find
a sh program, the PATH can be changed to find a compliant sh
to use. On platforms where /bin/sh is compliant, env should work
just fine.

git-svn-id: svn+ssh://svn.code.sf.net/p/check/code/trunk@1145 64e312b2-a51f-0410-8e61-82d0ca0eb02a