Do not vsnprintf expression string passed to ck_abort()
If ck_abort() detects a failure, the expression which was
evaluated is passed to _ck_assert_failed to print.
However, the previous behavior was to send the expression
to vsnprintf(), which would expect it to be properly formatted.

If the expression in ck_abort() contained % characters, such as:
   ck_abort(bar%foo == 0)
the "%f" portion would be printed as some non-existent floating
point number.

Now, if there is no message passed (and the expression is to
be used) simply report the expression as is. Only format
something with vsnprintf() if there is actual data that
expects formatting.

git-svn-id: svn+ssh://svn.code.sf.net/p/check/code/trunk@1183 64e312b2-a51f-0410-8e61-82d0ca0eb02a