checkmk: use double slash in regular expression
For many awk implementations using both a single and double slash
escapes in regex expressions in strings works correctly. However,
it was observed on Solaris that a single slash results in the
literal value being added to the string (e.g. \n was a newline)
and only the double slash works as expected.

As there should be no harm, updating the regex expressions to use
double slashes.

git-svn-id: svn+ssh://svn.code.sf.net/p/check/code/trunk@1229 64e312b2-a51f-0410-8e61-82d0ca0eb02a