use printf binary directly when forming expected test output
It was observed on Lubuntu that the sh built-in print does not
escape hex as expected. For example, the following is expected:

$ printf "hi\x9there\n"
hi	there
$

However, on Lubuntu the built-in printf command in sh produces the
following:

$ sh -c 'printf "hi\x9there\n'
hi\x9there
$

To avoid this, use the printf binary, which env will find and
invoke.