Remove AC_CONFIG_AUX_DIR
This is half an issue report and half a pull request.

The reason for this is that while autoreconf -fi works for most of the
needed scripts it doesn't copy ltmain.sh into build-aux which results
the following ./configure failure:

    configure.ac:137: error: required file 'build-aux/ltmain.sh' not found

This patch, which is a similar one I've provided to cloog[0] which
allows ./configure to at least complete.

I'm not sure what you think since obviously the directory was used for
some reason, likely to reduce top-level directory clutter, but this at
least works for me.

0 https://groups.google.com/forum/?hl=en#!topic/cloog-development/28_M3Qk87-Q

Signed-off-by: Earnestly <zibeon@gmail.com>