Disallow a test case from being added to a suite twice
It was possible to add a test case to a suite more than
once. When the suite is freed later on the test case is
freed twice, causing a double free error.

Although is likely a bug from the test writer, this change
allows check to disallow double additions to prevent
double freeing a test case.

https://github.com/libcheck/check/issues/50