Disable SIGFPE tests for Cygwin until fixed
The test_fpe and test_mark_point tests currently fail on Cygwin.
Instead of the test process receiving a SIGFPE it instead fails
with a SIGSEGV. The reason is not yet know. However, attempting a
simple program that installs a SIGFPE handler then does a
raise(SIGFPE) does result in the handler being called.

This tests need to be enabled again once the reason for the failures
is resolved.

https://github.com/libcheck/check/issues/97