Add warning on floating point eq and ne assertions
The usefulness of the float/double/ldouble eq and ne macros is very limited. Comparing
the results of a floating point computation should be done with some tolerance, instead
of expecting an exact answer. There may be differences on different platforms, etc.

To this end, some warnings are added to the eq and ne functions, instead directing users
to use the _tol functions instead.