Move CloudBees tests to Travis-CI
Earlier this year GitHub deprecated some encryption types from
accessing repos. This broke the FC17 executors on CloudBees.
Trying to update to using the FC25 executors found that not
all tools were installed. CloudBees determined it would take
too much time to install the missing tools on FC25 which were
on FC17 (for a FOSS project), and declined to do so. As a result,
Check's tests no longer run on CloudBees.

In response, the tests which ran on CloudBees are being moved
to Travis-CI. This includes the pre release check, building the
documentation, building with scan-build, and building with some
of the configure arguments.