Set  CMAKE_SOURCE_DIR for AppVeyor builds
Prior to CMake 3.13.3 if the source and bin directories were not specified it
was assumed to be the current working directory. Starting in 3.13.3 this case
results in an error. Neither was being specified in the AppVeyor builds. These
builds are now failing, and it is assumed to be related to a CMake upgrade.

Specifying the source dir should avoid the issue.

Reference: https://gitlab.kitware.com/cmake/cmake/issues/18817