Add libcompat code directly to libcheck
We have tried to statically link the built libcompat into built libcheck
in CMake without success. With autotools and libtool/similar this is
most likely very easy but no support on Windows.
So instead of building libcompat we incorporate the necessary parts
of libcompat code directly into libcheck when building libcheck.

Signed-off-by: Mikko Johannes Koivunalho <mikko.koivunalho@inoviagroup.se>