Add libcompat code directly to libcheck
We have tried to statically link the built libcompat into built libcheck
in CMake without success. With autotools (libtool et.al.) this is
most likely very easy but there is no support on Windows.
So instead of building libcompat we incorporate the necessary parts
of libcompat code directly into libcheck when building libcheck.

We would like to create an OBJECT library but currently they are
too unreliable and cumbersome,
especially with target_link_libraries and install(EXPORT...
https://stackoverflow.com/questions/38832528/transitive-target-include-directories-on-object-libraries

Signed-off-by: Mikko Johannes Koivunalho <mikko.koivunalho@iki.fi>