Remove non-CMake style option CHECK_ENABLE_TESTS
It is a CMake best practice to include file 'CTest'
in the project main file and then use option
'BUILD_TESTING' (defined in 'CTest') to enable or disable
building the tests. BUILD_TESTING is ON by default.

https://cmake.org/cmake/help/v3.9/module/CTest.html?highlight=build_testing
https://cmake.org/cmake/help/v3.9/command/add_test.html?highlight=build_testing

Option CHECK_ENABLE_TESTS is deprecated.
Because it is on by default, deprecation warning is only shown
to user when user explicitly turns CHECK_ENABLE_TESTS off.

Suggesting to remove the option by release 0.15.0.

Signed-off-by: Mikko Johannes Koivunalho <mikko.koivunalho@iki.fi>