Add param @ONLY to CMake configure_file()
Restrict variable replacement to references of the form @VAR@.
This is useful for configuring scripts that also use ${VAR} syntax.
Adding @ONLY is precautionary action to prevent future bugs.

Attn. cmake/config.h.in contains ${VAR} in #cmakedefine clauses!

How to use CMake variable substitution.

Wrong:

    #if @HAVE_FORK@
        [..]
    #endif

Because this will result as:

    #if 1
        [..]
    #endif

Better:

    /* Imported CMake variables created during build. */
    #cmakedefine HAVE_FORK 1

    #ifdef HAVE_FORK
        [..]
    #endif

Because this will result as:

    /* Imported CMake variables created during build. */
    #define HAVE_FORK 1

    #ifdef HAVE_FORK
        [..]
    #endif

Or:

    /* Imported CMake variables created during build. */
    /* #undef HAVE_FORK */
    [..]

This is more readable. Now we know which decision actually
was taken.

If you need to pass the parameter to the definition,
it is just as easy:

    /* Imported CMake variables created during build. */
    #cmakedefine CORRECT_SIZE @CORRECT_SIZE@

    #if CORRECT_SIZE >= 20
        [..]
    #endif

Signed-off-by: Mikko Johannes Koivunalho <mikko.koivunalho@iki.fi>