configure: check for makeinfo instead of tex for documentation
While upgrading the Alpine Linux check package to 0.13.0 I noticed that
documentation files were no longer installed, even though they were
installed successfully previously.

After investigating this for a while I noticed that this seems to be due
to an incorrect fix for #206. In #206 the issue reporter complains that
the build fails if texinfo is not installed, the fix in #212 checks for
tex instead of makeinfo though. makeinfo is the command used by the
implicit gnu make rule to generate info pages.