Change CMAKE_BINARY_DIR to CMAKE_CURRENT_BINARY_DIR
https://github.com/libcheck/check/issues/238
Getting things to work with CMake's FetchContent

FetchContent builds Check as a sub project, not as an external
project which is installed to a subdirectory and then
the installation is used.
In a sub project the variables CMAKE_BINARY_DIR and CMAKE_SOURCE_DIR
become dangerous as they now point to the main project's root directory.

We saw a similar case in https://github.com/libcheck/check/issues/194
wrong usage of CMAKE_PROJECT_NAME when embedding check