Add option to disable timeout unit tests in CMake
There was an option to enable/disable Check's timeout unit tests with autotools.
There was no such option for CMake, and the timeout tests were run unconditionally.

On some systems, such as overloaded computers or VMs, some of the timeout tests
which have smaller timeout values sporadically fail. The option to disable the timeout
tests on such machines will be important to reduce false positives.