Disable timeout tests on OSX for Action
The timeout tests have a few tests that use quick
timeouts to check correctness. On machines that are
overloaded or are shared these timeouts may be missed,
causing the tests to fail sporadically.

This commit removes the timeout tests from running on
OSX machines in GitHub Actions, as they sporadically
fail on the quick timeout values. This issue has not been
seen on the Linux Action machines yet, and will stay 
enabled.