Let fail_if bypass printf validation checks
There are some uses of fail_if in libcheck users which do not
supply an argument. fail_if() expects an argument, though when
none is provided it is ignored.

Check 0.15.0 (and a fix in 0.15.1) added printf argument validation
to the assert macros. As a result, instances of fail_if with
no arguments are now being flagged.

To support fail_if() without arguments, a new _ck_assert_failed*
method is being added whose purpose is to act just like
_ck_assert_failure but no printf validation is done on its
arguments by the compiler.