Fix GitHub Pages Site Download Bug #10
As the related issue documents, GitHub pages cannot
determine the file type of the NEWS or LGPL files because
they do not have a file extension, so the server uses a
default MIME type of application/octet-stream, which
triggers a download, rather than letting the user actually
read the file.

To circumvent this problem, I changed to file links from
relative links to full hyperlinks referencing the GitHub
repository URL instead. The NEWS file is now viewable via
the "raw" view option accessible through GitHub, and the
LGPL file links to the GitHub license file default.

I chose not to simply use the raw view for the license file
since I felt most people looking at that file would
probably not be lawyers, and would therefore benefit from
the summary GitHub displays at the top of the page.

This pull request fixes issue #10