Fixed CMake man page installation bug
The man pages need to be prefixed with
${CMAKE_CURRENT_BINARY_DIR} in order for CMake to actually
look in the configuration directory as opposed to the
directory where they were originally located.

I'm not sure how I missed this bug the first go-around, but
since the man pages are optional, CMake silently moves on
if it can't find the man pages where it thinks they should
be.

This bug has now been fixed (hopefully).