Removed Timestamps from Project Configuration Process
As discussed in the pull request (PR #310), the addition of
timestamps to the project configuration, and transitively
to the man pages, we sacrificed build reproducibility
without really getting much in return.

Therefore, this pull request removes the timestamp function
from the CMakeLists.txt file which contained it, as well as
the corresponding Autoconf macro file and the code that
called it from the Autotools build system.

Furthermore, the man pages' date section has been replaced
with an empty string, but the project version is still
configured via either CMake or Autotools onto the man pages
so that the reader is aware of the library version from
which they originated.