Enable Doxygen Documentation Generation
This pull request enables the automatic generation of
Doxygen HTML documentation via the check-doxygen-docs build
target.

This build target is not included in the Check library
default build list, as requested in issue #217, but instead
allows the user to specifically build this documentation
after project generation with the following command.

 $ cmake --build . --target check-doxygen-docs

The generated documentation is created in the project build
directory, in the doc/Doxygen folder.

In order to enable Doxygen, the project must be configured
using the -DCHECK_ENABLE_DOXYGEN_DOCS CMake flag.