Allow disabling autodetection of subunit library
It can be useful to avoid linking to subunit when we are building the check
library for the host, e.g. in a buildroot recipe, where the built check is
linked into other applications which are used on the host to build images.

These applications are built with the host's cross-compiler and can use
the host's pkgconfig to determine if libraries are available. When check
is linked against other libraries, it can fail this check for subunit.

Allow disabling the autodetection of subunit with a configure flag
-DENABLE_SUBUNIT_EXT=OFF

Signed-off-by: Christopher Obbard <christopher.obbard@linaro.org>